feat: :sparkles: The ambient medium can now be set for each analyzer individually.
