+++ opossum_gui/src/components/inputs/mod.rs
@@ -1,4 +1,6 @@
 pub mod absorption_editor;
+pub mod material_selector;
 pub mod refractive_index_editor;
 
+pub use material_selector::MaterialSelector;
 pub use refractive_index_editor::RefractiveIndexEditor;
